Remove middleware-navs.js (optimization)
It does redundant requests since there is a board list that always comes
with thread or board reply, as it has been requested per
U-Me-Chan/umechan#40.

It may cause some bugs though like displaying an empty boards list in
the navbar, but it actually happens before the boards list have been
fetched for the first time and it always fetches at home, board and
thread views, which are the most often visited views in this frontend.

But, let's see how it goes. Maybe it will work horribly byggy in prod
and I will have to redo this the other way with fallback requests
/boards/all to obtain the boards list.